requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Login
psycopg2
python-dotenv==0.11.0
# ----------------------
# Production Server
# ----------------------
gunicorn
